#52 feat : 회의록 단일조회 & 삭제 로직 변경

+++ src/main/java/com/example/myrok/service/RecordServiceImpl.java
@@ -2,6 +2,7 @@ package com.example.myrok.service;
 
 import com.example.myrok.domain.*;
 import com.example.myrok.domain.Record;
+import com.example.myrok.dto.MemberDto;
 import com.example.myrok.dto.RecordDTO;
 import com.example.myrok.dto.RecordResponseDTO;
 import com.example.myrok.dto.RecordUpdateDTO;
@@ -41,6 +42,10 @@ public class RecordServiceImpl implements RecordService{
     private MemberRecordService memberRecordService;
     @Autowired
     private MemberRecordRepository memberRecordRepository;
+    @Autowired
+    private RecordTagRepository recordTagRepository;
+    @Autowired
+    private MemberRepository memberRepository;
 
 
     @Transactional(rollbackFor = Exception.class)
@@ -86,12 +91,20 @@ public class RecordServiceImpl implements RecordService{
 
     @Transactional(rollbackFor = Exception.class)
     @Override
-    public void deleteUpdate(Long recordId) {
+    public void deleteUpdate(Long recordId, Long recordWriterId) {
         // 회의록 삭제
         Record record = recordRepository.findById(recordId)
                 .orElseThrow(() -> new EntityNotFoundException("존재하지 않는 회의록입니다. id: " + recordId));
+        // 이미 삭제된 회의록이면 예외
         if (record.getDeleted()) {
             throw new CustomException(ErrorCode.DELETED_RECORD_CODE, HttpStatus.BAD_REQUEST);
+        }
+        //회의 참여자가 아니라면 예외
+        MemberRecord memberRecord =memberRecordRepository.findByMemberIdAndRecordIdAndDeletedIsFalse(recordWriterId,recordId)
+                .orElseThrow(() -> new CustomException(ErrorCode.WRONG_RECORD_ACCESS, HttpStatus.BAD_REQUEST));
+        //작성자 외 삭제 시도시 예외
+        if (memberRecord.getRole()!=Role.ADMIN){
+            throw new CustomException(ErrorCode.WRONG_DELETE_ACCESS, HttpStatus.BAD_REQUEST);
         }
         record.delete();
         recordRepository.save(record);
@@ -136,6 +149,39 @@ public class RecordServiceImpl implements RecordService{
         record.setRecordTagList(updateRecordTagList);
 
         return recordRepository.save(record);
+    }
+
+    @Override
+    public RecordResponseDTO read(Long recordId){
+        Record record = recordRepository.findById(recordId)
+                .orElseThrow(() -> new EntityNotFoundException("존재하지 않는 회의록입니다. id: " + recordId));
+        if (record.getDeleted()) {
+            throw new CustomException(ErrorCode.DELETED_RECORD_CODE, HttpStatus.BAD_REQUEST);
+        }
+        List<String> tagList = recordTagRepository.findTagsByRecordIdAndDeletedIsFalse(recordId);
+        List<MemberDto.MemberNameDto> memberList = new ArrayList<>();
+        List<Long> members = memberRecordRepository.findMemberIdByRecordIdAndDeletedIsFalse(recordId);
+        for (Long memberId : members) {
+            String memberName = memberRepository.findNameById(memberId);
+            MemberDto.MemberNameDto member = MemberDto.MemberNameDto.builder()
+                    .name(memberName)
+                    .memberId(memberId)
+                    .build();
+            memberList.add(member);
+        }
+        RecordResponseDTO readRecord = RecordResponseDTO.builder()
+                .recordId(record.getId())
+                .recordName(record.getRecordName())
+                .recordDate(record.getRecordDate())
+                .recordWriterId(record.getRecordWriterId())
+                .recordContent(record.getRecordContent())
+                .tagList(tagList)
+                .memberList(memberList)
+                .tagList(tagList)
+                .build();
+
+        return readRecord;
+
     }
